requirements.txt

Flask==2.0.2
PyYAML
python-dotenv==0.19.2